Fixed integer overflow when using large meshes

+++ src/HM/RotationMap.cpp
@@ -26,7 +26,8 @@ vfps::RotationMap::RotationMap(PhaseSpace* in, PhaseSpace* out,
 							   const InterpolationType it,
 							   const RotationCoordinates rt,
 							   bool interpol_saturating) :
-	HeritageMap(in,out,xsize,ysize,xsize*ysize*it*it/ROTMAP_SIZE,it*it,it),
+	HeritageMap(in,out,xsize,ysize,
+				size_t(xsize)*size_t(ysize)*it*it/ROTMAP_SIZE,it*it,it),
 	_sat(interpol_saturating)
 {
 	// gridpoint matrix used for interpolation
